IMCMS-335 New design to super admin page: categories tab - Change update and delete

+++ src/main/java/com/imcode/imcms/domain/service/api/DefaultCategoryService.java
@@ -3,9 +3,11 @@ package com.imcode.imcms.domain.service.api;
 import com.imcode.imcms.domain.dto.CategoryDTO;
 import com.imcode.imcms.domain.service.CategoryService;
 import com.imcode.imcms.model.Category;
+import com.imcode.imcms.model.CategoryType;
 import com.imcode.imcms.persistence.entity.CategoryJPA;
 import com.imcode.imcms.persistence.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -44,8 +46,39 @@ class DefaultCategoryService implements CategoryService {
     }
 
     @Override
+    public Category update(Category updateMe) {
+        Integer id = updateMe.getId();
+        String name = updateMe.getName();
+        CategoryType categoryType = updateMe.getType();
+        String description = updateMe.getDescription();
+        String imageUrl = updateMe.getImageUrl();
+
+        Category receivedCategory = categoryRepository.findOne(id);
+
+        for (Category category : getAll()) {
+            if (receivedCategory.getName().equals(category.getName())) {
+                throw new IllegalArgumentException();
+            } else if (receivedCategory.getName().isEmpty()) {
+                throw new IllegalArgumentException();
+            } else {
+                receivedCategory.setName(name);
+                receivedCategory.setType(categoryType);
+                receivedCategory.setDescription(description);
+                receivedCategory.setImageUrl(imageUrl);
+            }
+        }
+        final Category updatedCategory = categoryRepository.save(new CategoryJPA(receivedCategory));
+        return new CategoryDTO(updatedCategory);
+    }
+
+    @Override
     public void delete(int id) {
-        categoryRepository.deleteDocumentCategory(id);
-        categoryRepository.delete(id);
+        List<Integer> categoryDocIds = categoryRepository.findCategoryDocIds(id);
+        if (categoryDocIds.isEmpty()) {
+            categoryRepository.deleteDocumentCategory(id);
+            categoryRepository.delete(id);
+        } else {
+            throw new EmptyResultDataAccessException(id);
+        }
     }
 }
